chore: add enum feature
