+++ src/compiler_v/lexer/get_functions.v
@@ -110,7 +110,6 @@ fn (mut l Lexer) get_text_delim(kind token.Kind, delim_start string, delim_end s
 		str := l.input[start_pos..(l.pos - delim_end.len)].bytestr().replace('\n', '\\n')
 		return l.new_token(str, kind, 0)
 	} else {
-		println('ignore ${l.input[l.pos]}')
 		return l.new_token('', .ignore, 1)
 	}
 }
